+++ src/app/wallets/wallet-list-header/wallet-list-header.component.spec.ts
@@ -1,7 +1,7 @@
 import { byTestId, createHostFactory, SpectatorHost } from "@ngneat/spectator";
 
 import { WalletListActionsComponentModule } from "../wallet-list-actions/wallet-list-actions.component.module";
-import { WalletListHeaderComponentModule } from "../wallet-list-actions/wallet-list-header.component.module";
+import { WalletListHeaderComponentModule } from "../wallet-list-header/wallet-list-header.component.module";
 import { WalletListHeaderComponent } from "./wallet-list-header.component";
 
 describe("Wallet list header", () => {
