+++ src/app/wallets/wallet-list/wallet-list.component.html
@@ -1,4 +1,7 @@
-<div class="flex flex-col items-center mt-20 c-wallets__list">
+<div
+	class="flex flex-col items-center mt-20 wallet-list"
+	data-testid="wallet-list"
+>
 	<wallet-list-header
 		[name]="name"
 		[orientation]="headerOrientation"
