+++ src/app/wallets/wallet-list/wallet-list.component.spec.ts
@@ -1,63 +1,34 @@
-import {
-	ActionSheetController,
-	IonicModule,
-	NavController,
-} from "@ionic/angular";
-import {
-	byTestId,
-	createHostFactory,
-	createSpyObject,
-	mockProvider,
-	SpectatorHost,
-} from "@ngneat/spectator";
+import { IonicModule } from "@ionic/angular";
+import { byTestId, createHostFactory, SpectatorHost } from "@ngneat/spectator";
 import { TranslateModule } from "@ngx-translate/core";
 
-import { UserDataProviderMock } from "@@/test/mocks";
-import { AuthController } from "@/app/auth/shared/auth.controller";
-import { ArkApiProvider } from "@/services/ark-api/ark-api";
-import { MarketDataProvider } from "@/services/market-data/market-data";
-import { SettingsDataProvider } from "@/services/settings-data/settings-data";
-import { UserDataService } from "@/services/user-data/user-data.interface";
-import ApiClient from "@/utils/ark-client";
+import { WalletListHeaderComponentModule } from "../wallet-list-header/wallet-list-header.component.module";
+import { WalletListComponent } from "./wallet-list.component";
 
-import { WalletsEmptyListComponentModule } from "./wallet-list-empty/wallet-list-empty.module";
-import { WalletsComponent } from "./wallets.component";
-
-describe("Wallets Component", () => {
-	let spectator: SpectatorHost<WalletsComponent>;
-
-	const apiClient = createSpyObject(ApiClient);
+describe("Wallet List Component", () => {
+	let spectator: SpectatorHost<WalletListComponent>;
 
 	const createHost = createHostFactory({
-		component: WalletsComponent,
-		mocks: [AuthController, NavController, ActionSheetController],
+		component: WalletListComponent,
 		imports: [
 			IonicModule.forRoot(),
 			TranslateModule.forRoot(),
-			WalletsEmptyListComponentModule,
-		],
-		providers: [
-			{ provide: UserDataService, useClass: UserDataProviderMock },
-			mockProvider(MarketDataProvider),
-			mockProvider(SettingsDataProvider),
-			mockProvider(ArkApiProvider, {
-				client: apiClient,
-			}),
+			WalletListHeaderComponentModule,
 		],
 	});
 
 	it("should create", () => {
 		spectator = createHost(
-			`<ion-content><wallets-component></wallets-component></ion-content>`,
+			`<ion-content><wallet-list></wallet-list></ion-content>`,
 		);
 		expect(spectator.component).toBeTruthy();
 	});
 
-	it("should render the empty list", () => {
+	it("should render the list header", () => {
 		spectator = createHost(
-			`<ion-content><wallets-component></wallets-component></ion-content>`,
+			`<ion-content><wallet-list></wallet-list></ion-content>`,
 		);
 
-		expect(spectator.query(byTestId("c-wallets__empty-list"))).toBeTruthy();
+		expect(spectator.query(byTestId("wallet-list-header"))).toBeTruthy();
 	});
 });
